refactor: move ClientOnly to components
